[NFC] Split PaddingClearer out into a CodeGenUtils shared directory
We have plenty of stuff that needs to be shared between Classic-Codegen
and CIR, that we've been copy/pasting too often. This patch
proposes/produces a shared directory of CodeGenUtils that will be for
things that don't really belong elsewhere.  This functionality
(calculating padding for __builtin_clear_padding) is both too complex to
copy/paste, and doesn't really have reason to live anywhere except
CodeGen/CIR-CodeGen.

The type has been split out as faithfully as possible, with only minor
changes to make it LLVM-agnostic.  The part of it that is required to
generate the LLVM-IR is left in CGBuiltin.cpp, but takes the list of
padding-intervals as an input, which will be useful for CIR's
representation.
